[aptos-framework] An example token transfer harness
In Move, it is dangerous to support unilateral movement of resources,
because it could result in an account accruing data and increasing the
cost of operating said account.

Tokens must be offered to retrieval and then claimed in a two phase
process. The TokenTransfers shows a method for doing so. The same
harness can be used to build auctions and store fronts.