[framework] split up initializing validator config
A custodian likely will not generate a valid validator config. In this
commit, we stop coercing them into generating a valid bls key just to
register and ultimately have it being wrong. Instead they can register
the owner only status, pass on the operator capability, and have that
entity rotate in a key triggering that their account is now valid.