[aptos-cli] Pull latest packages by default for cached branch dependencies
Now all the confusion is gone around why dependencies are out of date,
it follows similar to cargo, where you want to use a git hash if you
want frozen dependencies.  Otherwise, it will updated.  Additionally,
there's a flag to turn it off if your network is disconnected, or
something like that.