[move] Use workspace deps for Move and upgrade to Rust 1.64 (#5113)
This uses the new workspace dep feature to organize Move dependencies. The version is specified in the workspace Cargo.toml, and a crate can just use `move-core-types = { workspace = true }`. Features are additative, solving the unwanted feature unification problem. The `move-deps` create has been removed for good, which should make individual compilation faster. (In the majority of cases, aptos crates only need `move-core-types`.) For developing against a local Move copy, one can just swap all the references in the workspace Cargo.toml.

This requires an update to 1.64 and a couple of new clippy errors have been fixed.

I would argue this is safe both to pick or to not pick into release branches. Though a lot of files changed, the change is localized to some imports and part of toml. The reason not to pick is that we don't think it is safe to switch release branches to Rust 1.64.

Notice that the nicer syntax `move-core-types.workspace = true` doesn't work as it crashes `cargo sort`.