deprecate stale fee collection implementation (#12800)
Current implementation is stale, and cannot be used as defined. So better to remove it, and add it back later if needed.

Let me know if you have preference on whether to keep or remove stake related changes, as those might be more similar for future stake changes. Removed them in this PR, as adding them back should be as easy as this removal is.