[function values] Mark language version 2.2 to be stable and bytecode version 8 the default (#17116)
This configures 2.2 which enables function values to be stable on compiler side, allowing people to submit to chain without errors or warnings. Also, the bytecode version is to have default of VERSION_8. This are all changes on CLI/compiler side and not for node code, which is guarded by according feature flags.